Use full URL in repository blacklist
This creates consistency with the hardware package list, which I plan to make a comprehensive list of all hardware packages, regardless of whether they're hosted on GitHub.

It makes it easier to copy/paste the repository URLs to/from the browser.

It enables the (unlikely) possibility of someday adding support for other hosts to the project.