Split the CI build into multiple jobs
Although this likely will result in a slightly slower CI build due to preparing a separate virtual machine for each job, it makes it so that a summary of the build results can be seen just by checking the job list. It also avoids any possibility of one test interfering with the others.